Bug 583652 - Console caches magic numbers instead of DNA-generated values
The users and groups tab currently fills in the contents of an entry after
an add or modify using the values one entered into Console.  If the server
added any other attributes or modified any values (such as DNA does), those
values will not be shown unless one searches for the entry again.

This patch causes the Consoel to refresh the entry by searching for it
again after an add or modify operation.  The result is that one sees the
actual contents of the entry as they exist in the server.