Bug: 594939 - ACI editing dialog initial size is not big enough to display
The problem had to do with the preferred size of the JTextArea that holds the
ACI name having an unusually large height.  The JTextArea had line wrapping
enabled, yet it's default size was not set.  This was resulting in it thinking
that it needed a huge height due to a lot of text wrapping.

The fix is to set the initial rows and columns of the text area when we call
it's constructor and to contain the text area in a scrollable pane.  This
results in a sane preferred size and allows a long ACI name to line wrap
properly.