Bug 654990 - Certificate prompt window does not autosize
The JTextArea that we use in the dialog that prompts for certificate trust is
unable to determine it's proper preferred size since it is created with no rows
and no columns.  This requires us to call pack() twice before it can get it's
proper preferred size.  This is detailed in
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=4446522.