Bug 229699 - objectclass without parent causes StringIndexOutOfBounds in console
The exception was occuring since the Console assumed that an entry created by
going to "New->User" would have the givenname and sn attribute present and
never checked if they were null before calling the String.substring() method.
The fix is to first check if these attributes are null.

There were other issues around this problem as well.  The
ResourcePageObservable class that is used to represent the entry being edited
keeps a list of objectclasses along with the list of attributes and values for
the entry (including the objectclass attribute).  The problem was that removing
an objectclass from the entry in the editor did not remove the objectclass from
the objectclass list.  This causes the editor to make assumptions about the
entry, such as the presence of the givenname and sn attribute that caused the
exception.  The fix is to sync the objectclass list with the actual
objectclasses that are present in the entry when a change is made.