Bug 158926 - Unable to install CA certificate when using hardware token              ( LunaSA )
https://bugzilla.redhat.com/show_bug.cgi?id=158926

Description: CertificateDialog passes the token name selected from the
Security Device menu to ServerCertificatePane, but NOT to CACertificatePane.
Due to this, when a hardware token was selected as a Security Device on
CACertificatePane, installing CA cert was not forwarded to the hardware
token.  Also, it was missing to send a passowrd for the Security Device on
CACertificatePane, which caused the install fail with an error "Password".

This patch solves the problems.