Bug 622436 - Removal of Security:domestic from Console
This patch removes the "Security level" field from the server info panels in
Console.  One will no longer see this field.

Internally, the Console code still has a security level concept that is used
for enabling ciphers.  It is not worth changing this as the code is widespread
and exposed via public API.  It is possible that it is used by other projects,
such as Dogtag.