Bug 706258 - ACI Editor dialog flickers
The ACI editor dialog draws itself too large for the screen when it is
launched in manual edit mode.  This causes it to continually redraw itself,
which makes it flicker on screen.  This is similar to the previous issue
fixed as bug 594939, but the difference is that the editor needs to be
launched in manual mode, not switched from visual mode.

The fix is to set a default size for the text area used to display the
ACI value.  This causes the JTextArea preferred component size to be set
correctly, which results in the window being drawn at the proper size.