Bug 706472 - [console] java exception throw in UI, but user gets created
https://bugzilla.redhat.com/show_bug.cgi?id=706472
Resolves: bug 706472
Bug Description: [console] java exception throw in UI, but user gets created
Reviewed by: nkinder (Thanks!)
Branch: master
Fix Description: This problem is caused when you add a new entry to a
read-only replica.  The console configures the ldapjdk to automatically
follow referrals back to one of the masters, so the add operation is
successful.  The console then does a refresh so it can get the newly added
entry with any automatic attributes added, such as memberof or uidnumber.
If this is a read-only replica, the entry won't be available immediately.
The fix is to keep trying to read for 10 seconds.  If the entry is not
available after that time, put up a dialog telling the user to Refresh
until the entry is available.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no