Ticket 362 - Directory Console generates insufficient key strength
The Console certificate request wizard offers no control over the key
size and signing algorithm that is used.  The security CGI that the
Console calls supports different key sizes and signing algorithms
now, so we can expose this capability in the Console UI.

This patch adds an additional page to the certificate request wizard.
This page has two combo boxes that allow the RSA key size and the
signing algorithm to be selected.  The possible values are defined in
a properties file, which will make changes easy in the future as long
as the security CGI allows the new values.  The new RSA key size
values are 2048, 3072, and 4096 with the default selection of 2048.
The new signing algorithm values are SHA-1, SHA-256, SHA-384, and
SHA-512 with the default selection of SHA-1.