Ticket 47467 - Improve CRL import dialog text
The CRL/CKL import dialog doesn't give any information about the
required format for the CRL/CKL file.  This patch improves the text
to state that the file should be in PEM format and it needs to exist
in the server security directory.  The message was also tweaked
slightly to indicate that a filename should be specified as opposed
to a path.  I didn't want to make the message too long,
but this at least provides a hint.  If a path is entered, the
security CGI returns a detailed error messages stating that full and
absolute paths are not allowed as well as the expected file location.