Ticket #48811 - Console window could be hidden after login via consoles on multiple hosts
Bug Description: When Console is displayed, it uses the values from the
preference to determine the location and the size.  The preference is
stored under o=netscaperoot in the config ldap server in the centralized
manner.  But the values depend upon the window system resolution.  That
means the location which is good on a window system may not be valid on
others.

Fix Description: Compare the location from the preference with the display
size.  If the stored location is outside of the display, reset it to the
centre.

https://fedorahosted.org/389/ticket/48811

Tested and reviewed by vashirov@redhat.com and wibrown@redhat.com.
(Thank you, Viktor and William!!)